fix: Removed openPacket and moved pcap tests
