+++ src/js/brim/ingest/detectFileTypes.test.js
@@ -1,15 +1,13 @@
 /* @flow */
 
-import path from "path"
-
 import ingest from "./"
+import itestFile from "../../test/itestFile"
 
-const testFile = (name) => path.join(__dirname, "../../../../itest/", name)
-const pcap = testFile("sample.pcap")
-const pcapng = testFile("sample.pcapng")
-const zeek = testFile("sample.tsv")
-const json = testFile("sample.ndjson")
-const unknown = testFile("setup.js")
+const json = itestFile("sample.ndjson")
+const pcap = itestFile("sample.pcap")
+const pcapng = itestFile("sample.pcapng")
+const unknown = itestFile("setup.js")
+const zeek = itestFile("sample.tsv")
 
 test("add file types", async () => {
   let paths = [pcap, pcapng, zeek, json, unknown]
