+++ src/js/test/itestFile.js
@@ -1,0 +1,7 @@
+/* @flow */
+
+import path from "path"
+
+export default (name: string) => {
+  return path.join(__dirname, "../../../itest", name)
+}
